# backend/api-document.json
{"openapi":"3.0.0","paths":{"/users":{"get":{"operationId":"UsersController_getUsers","parameters":[{"name":"id","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["users"]},"post":{"operationId":"UsersController_createUsers","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["users"]}},"/users/db":{"get":{"operationId":"UsersController_getAllUsersFromDB","parameters":[],"responses":{"200":{"description":""}},"tags":["users"]},"post":{"operationId":"UsersController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["users"]}},"/users/db/{id}":{"get":{"operationId":"UsersController_getUsersByIdFromDB","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["users"]},"put":{"operationId":"UsersController_replaceUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplaceUserDto"}}}},"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]},"patch":{"operationId":"UsersController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UsersController_deleteUsersFromDB","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}},"/users/{id}":{"get":{"operationId":"UsersController_getUserById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["users"]},"put":{"operationId":"UsersController_updateUsers","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplaceUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["users"]},"patch":{"operationId":"UsersController_updateUsersPatch","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["users"]},"delete":{"operationId":"UsersController_deleteUsers","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["users"]}},"/posts":{"get":{"operationId":"PostsController_getAllPosts","parameters":[{"name":"userId","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["posts"]},"post":{"operationId":"PostsController_createPost","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostDTO"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["posts"]}},"/posts/db":{"get":{"operationId":"PostsController_getAllPostsFromDB","parameters":[],"responses":{"200":{"description":""}},"tags":["posts"]},"post":{"operationId":"PostsController_createPostDB","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostDTO"}}}},"responses":{"201":{"description":""}},"tags":["posts"],"security":[{"bearer":[]}]}},"/posts/db/{id}":{"get":{"operationId":"PostsController_getPostByIdDB","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["posts"]},"put":{"operationId":"PostsController_replacePostDB","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplacePostDTO"}}}},"responses":{"200":{"description":""}},"tags":["posts"],"security":[{"bearer":[]}]},"patch":{"operationId":"PostsController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePostDTO"}}}},"responses":{"200":{"description":""}},"tags":["posts"],"security":[{"bearer":[]}]},"delete":{"operationId":"PostsController_deleteUsersFromDB","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["posts"],"security":[{"bearer":[]}]}},"/posts/db/post-with-user/{id}":{"get":{"operationId":"PostsController_getPostWithUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["posts"]}},"/posts/{id}":{"get":{"operationId":"PostsController_getPostById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["posts"]},"put":{"operationId":"PostsController_replacePost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReplacePostDTO"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["posts"]},"patch":{"operationId":"PostsController_updatePostPatch","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePostDTO"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["posts"]},"delete":{"operationId":"PostsController_deletePost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["posts"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthEntity"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthEntity"}}}}},"tags":["auth"]}},"/auth/profile":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["auth"]}}},"info":{"title":"API documentation","description":"The API description","version":"0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"GeoDto":{"type":"object","properties":{"lat":{"type":"string"},"lng":{"type":"string"}},"required":["lat","lng"]},"AddressDto":{"type":"object","properties":{"street":{"type":"string"},"suite":{"type":"string"},"city":{"type":"string"},"zipcode":{"type":"string"},"geo":{"$ref":"#/components/schemas/GeoDto"}},"required":["street","suite","city","zipcode","geo"]},"CompanyDto":{"type":"object","properties":{"name":{"type":"string"},"catchPhrase":{"type":"string"},"bs":{"type":"string"}},"required":["name","catchPhrase","bs"]},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"address":{"$ref":"#/components/schemas/AddressDto"},"phone":{"type":"string"},"website":{"type":"string"},"company":{"$ref":"#/components/schemas/CompanyDto"}},"required":["name","username","password","email","address","phone","website","company"]},"ReplaceUserDto":{"type":"object","properties":{"name":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string","nullable":true,"default":null},"address":{"$ref":"#/components/schemas/AddressDto"},"phone":{"type":"string","nullable":true,"default":null},"website":{"type":"string","nullable":true,"default":null},"company":{"$ref":"#/components/schemas/CompanyDto"}},"required":["name","username","password"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"address":{"$ref":"#/components/schemas/AddressDto"},"phone":{"type":"string"},"website":{"type":"string"},"company":{"$ref":"#/components/schemas/CompanyDto"}}},"CreatePostDTO":{"type":"object","properties":{"userId":{"type":"number","nullable":true,"default":null},"title":{"type":"string","nullable":true,"default":null},"body":{"type":"string","nullable":true,"default":null}},"required":["userId","title","body"]},"ReplacePostDTO":{"type":"object","properties":{"userId":{"type":"number","nullable":true,"default":null},"title":{"type":"string","nullable":true,"default":null},"body":{"type":"string","nullable":true,"default":null}},"required":["userId","title","body"]},"UpdatePostDTO":{"type":"object","properties":{"userId":{"type":"number","nullable":true,"default":null},"title":{"type":"string","nullable":true,"default":null},"body":{"type":"string","nullable":true,"default":null}},"required":["userId","title","body"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string","minLength":6}},"required":["email","password"]},"AuthEntity":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]}}}}